Revert "Revert "Move Lua scripts from ts files into lua files in Corescripts directory""
